Starter refactoring
