+++ internal/app/starter/master.go
@@ -3,13 +3,7 @@
 // LICENSE.md file distributed with the sources of this project regarding your
 // rights to use or distribute this software.
 
-package main
-
-/*
-#include "c/starter.c"
-*/
-// #cgo CFLAGS: -I..
-import "C"
+package starter
 
 import (
 	"fmt"
@@ -20,22 +14,20 @@ import (
 	"runtime"
 	"syscall"
 	"time"
-	"unsafe"
 
 	"github.com/sylabs/singularity/internal/pkg/runtime/engines"
-	"github.com/sylabs/singularity/internal/pkg/runtime/engines/config/starter"
+	starterConfig "github.com/sylabs/singularity/internal/pkg/runtime/engines/config/starter"
 	"github.com/sylabs/singularity/internal/pkg/sylog"
-	"github.com/sylabs/singularity/internal/pkg/util/mainthread"
 )
 
-// SMaster initializes a runtime engine and runs it
-func SMaster(rpcSocket, masterSocket int, starterConfig *starter.Config, jsonBytes []byte) {
+// Master initializes a runtime engine and runs it
+func Master(rpcSocket, masterSocket int, sconfig *starterConfig.Config, jsonBytes []byte) {
 	var fatal error
 	var status syscall.WaitStatus
 
 	fatalChan := make(chan error, 1)
 	ppid := os.Getppid()
-	containerPid := starterConfig.GetContainerPid()
+	containerPid := sconfig.GetContainerPid()
 	engine, err := engines.NewEngine(jsonBytes)
 	if err != nil {
 		sylog.Fatalf("failed to initialize runtime: %s\n", err)
@@ -72,7 +64,7 @@ func SMaster(rpcSocket, masterSocket int, starterConfig *starter.Config, jsonByt
 
 		_, err = conn.Read(data)
 		if err != nil && err != io.EOF {
-			if starterConfig.GetInstance() && os.Getppid() == ppid {
+			if sconfig.GetInstance() && os.Getppid() == ppid {
 				syscall.Kill(ppid, syscall.SIGUSR2)
 			}
 			fatalChan <- fmt.Errorf("failed to start process: %s", err)
@@ -80,13 +72,13 @@ func SMaster(rpcSocket, masterSocket int, starterConfig *starter.Config, jsonByt
 		}
 		err = engine.PostStartProcess(containerPid)
 		if err != nil {
-			if starterConfig.GetInstance() && os.Getppid() == ppid {
+			if sconfig.GetInstance() && os.Getppid() == ppid {
 				syscall.Kill(ppid, syscall.SIGUSR2)
 			}
 			fatalChan <- fmt.Errorf("post start process failed: %s", err)
 			return
 		}
-		if starterConfig.GetInstance() {
+		if sconfig.GetInstance() {
 			// sleep a bit to see if child exit
 			time.Sleep(100 * time.Millisecond)
 			if os.Getppid() == ppid {
@@ -112,7 +104,7 @@ func SMaster(rpcSocket, masterSocket int, starterConfig *starter.Config, jsonByt
 	runtime.UnlockOSThread()
 
 	if fatal != nil {
-		if starterConfig.GetInstance() {
+		if sconfig.GetInstance() {
 			if os.Getppid() == ppid {
 				syscall.Kill(ppid, syscall.SIGUSR2)
 			}
@@ -129,7 +121,7 @@ func SMaster(rpcSocket, masterSocket int, starterConfig *starter.Config, jsonByt
 		syscall.Kill(syscall.Gettid(), syscall.SIGKILL)
 	} else if status.Exited() {
 		sylog.Debugf("Child exited with exit status %d", status.ExitStatus())
-		if starterConfig.GetInstance() {
+		if sconfig.GetInstance() {
 			if status.ExitStatus() != 0 {
 				if os.Getppid() == ppid {
 					syscall.Kill(ppid, syscall.SIGUSR2)
@@ -141,59 +133,5 @@ func SMaster(rpcSocket, masterSocket int, starterConfig *starter.Config, jsonByt
 			}
 		}
 		os.Exit(status.ExitStatus())
-	}
-}
-
-func startup() {
-	loglevel := os.Getenv("SINGULARITY_MESSAGELEVEL")
-	os.Clearenv()
-	if loglevel != "" {
-		if os.Setenv("SINGULARITY_MESSAGELEVEL", loglevel) != nil {
-			sylog.Warningf("can't restore SINGULARITY_MESSAGELEVEL environment variable")
-		}
-	}
-
-	cconf := unsafe.Pointer(&C.config)
-	starterConfig := starter.NewConfig(starter.CConfig(cconf))
-	jsonBytes := C.GoBytes(unsafe.Pointer(C.json_stdin), C.int(starterConfig.GetJSONConfSize()))
-
-	// free allocated buffer
-	C.free(unsafe.Pointer(C.json_stdin))
-	if unsafe.Pointer(C.nspath) != nil {
-		C.free(unsafe.Pointer(C.nspath))
-	}
-
-	switch C.execute {
-	case C.SCONTAINER_STAGE1:
-		sylog.Verbosef("Execute scontainer stage 1\n")
-		SContainer(int(C.SCONTAINER_STAGE1), int(C.master_socket[1]), starterConfig, jsonBytes)
-	case C.SCONTAINER_STAGE2:
-		sylog.Verbosef("Execute scontainer stage 2\n")
-		mainthread.Execute(func() {
-			SContainer(int(C.SCONTAINER_STAGE2), int(C.master_socket[1]), starterConfig, jsonBytes)
-		})
-	case C.SMASTER:
-		sylog.Verbosef("Execute smaster process\n")
-		SMaster(int(C.rpc_socket[0]), int(C.master_socket[0]), starterConfig, jsonBytes)
-	case C.RPC_SERVER:
-		sylog.Verbosef("Serve RPC requests\n")
-		RPCServer(int(C.rpc_socket[1]), C.GoString(C.sruntime))
-	}
-	sylog.Fatalf("You should not be there\n")
-}
-
-func init() {
-	// lock main thread for function execution loop
-	runtime.LockOSThread()
-	// this is mainly to reduce memory footprint
-	runtime.GOMAXPROCS(1)
-}
-
-func main() {
-	go startup()
-
-	// run functions requiring execution in main thread
-	for f := range mainthread.FuncChannel {
-		f()
 	}
 }
