+++ internal/app/starter/rpc.go
@@ -3,7 +3,7 @@
 // LICENSE.md file distributed with the sources of this project regarding your
 // rights to use or distribute this software.
 
-package main
+package starter
 
 import (
 	"net"
